refactor: TableVC -> CustomTableVC과 MainVC으로 나누는 중(tableView가 state를 가질수 있게, MassiveVC 피하기 위해)
